requirements.txt
anthropic>=0.7.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pytest>=7.0.0
pytest-cov>=4.0.0
